+++ R/plot.telemetry.R
@@ -11,7 +11,6 @@ zoom.telemetry <- function(x,fraction=1,...)
 methods::setMethod("zoom",signature(x="telemetry"), function(x,fraction=1,...) zoom.telemetry(x,fraction=fraction,...))
 methods::setMethod("zoom",signature(x="UD"), function(x,fraction=1,...) zoom.telemetry(x,fraction=fraction,...))
 
-
 ##############
 new.plot <- function(data=NULL,CTMM=NULL,UD=NULL,level.UD=0.95,level=0.95,units=TRUE,fraction=1,add=FALSE,xlim=NULL,ylim=NULL,ext=NULL,...)
 {
@@ -205,26 +204,10 @@ plot.telemetry <- function(x,CTMM=NULL,UD=NULL,level.UD=0.95,level=0.95,DF="CDF"
   #########################
   # PLOT TELEMETRY DATA
 
-  # prepare point characteristics
-  prepare.p <- function(pchar,all=FALSE)
-  {
-    if(!is.list(pchar))
-    {
-      if(length(x)>1)
-      {
-        pchar <- array(pchar,length(x))
-        pchar <- as.list(pchar)
-      }
-      else if(!all)
-      { pchar <- list(array(pchar,length(x[[1]]$t))) }
-    }
-    return(pchar)
-  }
-
-  col <- prepare.p(col)
-  pch <- prepare.p(pch)
-  lwd <- prepare.p(lwd)
-  type <- prepare.p(type,all=TRUE)
+  col <- format.par(col,x)
+  pch <- format.par(pch,x)
+  lwd <- format.par(lwd,x)
+  type <- format.par(type,x,all=TRUE)
   error <- rep(error,length(x))
 
   # automagic the plot point size
@@ -233,7 +216,7 @@ plot.telemetry <- function(x,CTMM=NULL,UD=NULL,level.UD=0.95,level=0.95,DF="CDF"
     p <- sum(sapply(x, function(d) { length(d$t) } ))
     if(p>1000) { cex <- 1000/p } else { cex <- 1 }
   }
-  cex <- prepare.p(cex)
+  cex <- format.par(cex,x)
 
   # standard deviations to plot for circle/ellipse
   z <- sqrt(-2*log(alpha.UD))
@@ -373,12 +356,37 @@ plot.telemetry <- function(x,CTMM=NULL,UD=NULL,level.UD=0.95,level=0.95,DF="CDF"
 #methods::setMethod("plot",signature(x="telemetry"), function(x,...) plot.telemetry(x,...))
 
 
+# format point characteristics for dataset x
+format.par <- function(pchar,x,all=FALSE)
+{
+  if(!is.list(pchar))
+  {
+    if(length(x)>1)
+    {
+      pchar <- array(pchar,length(x))
+      pchar <- as.list(pchar)
+    }
+    else if(!all)
+    { pchar <- list(array(pchar,length(x[[1]]$t))) }
+  }
+  return(pchar)
+}
+
+
+# get par that may be a constant or variable
+pull <- function(pchar,i)
+{
+  if(length(pchar)>1) { pchar <- pchar[i] }
+  return(pchar)
+}
+
+
 ##############
 plot.UD <- function(x,level.UD=0.95,level=0.95,DF="CDF",units=TRUE,col.level="black",col.DF="blue",col.grid="white",labels=NULL,fraction=1,add=FALSE,xlim=NULL,ylim=NULL,ext=NULL,cex=NULL,lwd=1,...)
 {
   x <- listify(x)
 
-  dist <- new.plot(UD=x,units=units,fraction=fraction,add=add,xlim=xlim,ylim=ylim,ext=ext,level.UD=level.UD,level=level,...)
+  dist <- new.plot(UD=x,units=units,fraction=fraction,add=add,xlim=xlim,ylim=ylim,ext=ext,level.UD=level.UD,level=level,cex=cex,...)
 
   # contours colour
   if(length(col.level)==length(level.UD) && length(col.level) != length(x))